backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.2
python-multipart==0.0.6
